AVRO-4295: [csharp] Use long for collection min-bytes to avoid int overflow
Review feedback: the map's minBytes was computed as 1 + MinBytesPerElement(...).
Since MinBytesPerElement clamps to int.MaxValue, adding 1 in int arithmetic
overflows to a negative value, which makes minBytesPerElement <= 0 and silently
disables the remaining-bytes validation for maps. Compute both the array and map
minima as long (1L + ...) and take a long in EnsureCollectionAvailable.

+++ lang/csharp/src/apache/main/Generic/GenericReader.cs
@@ -404,7 +404,7 @@ namespace Avro.Generic
             ArraySchema rs = (ArraySchema)readerSchema;
             object result = CreateArray(reuse, rs);
             int i = 0;
-            int minBytes = MinBytesPerElement(writerSchema.ItemSchema);
+            long minBytes = MinBytesPerElement(writerSchema.ItemSchema);
             for (long nl = d.ReadArrayStart(); nl != 0; nl = d.ReadArrayNext())
             {
                 // Reject a block whose element count could not be backed by the
@@ -497,7 +497,7 @@ namespace Avro.Generic
             MapSchema rs = (MapSchema)readerSchema;
             object result = CreateMap(reuse, rs);
             // Map keys are strings (>= 1 byte length prefix) plus the value.
-            int minBytes = 1 + MinBytesPerElement(writerSchema.ValueSchema);
+            long minBytes = 1L + MinBytesPerElement(writerSchema.ValueSchema);
             for (long nl = d.ReadMapStart(); nl != 0; nl = d.ReadMapNext())
             {
                 EnsureCollectionAvailable(d, nl, minBytes);
@@ -573,7 +573,7 @@ namespace Avro.Generic
         /// allocating. Skipped when the per-element minimum is zero, or when the
         /// decoder cannot report how many bytes remain.
         /// </summary>
-        private static void EnsureCollectionAvailable(Decoder d, long count, int minBytesPerElement)
+        private static void EnsureCollectionAvailable(Decoder d, long count, long minBytesPerElement)
         {
             if (count <= 0)
             {
